Repair character positions without touching the configuration
JSHint reports physical character positions instead of a character
index, i.e. all tab characters are multiplied with the (configurable)
indent. We used to set the option "indent" to 1 to get the real
character index. But this setting could be overwritten in a custom
configuration.

Now we leave the configuration untouched and repair the character
position based on the configured or default indent.