Handle exceptions caused by JSHint
So far, we only handled JavaScript errors, which are explicitly thrown
by JavaScript code. We did not handle exceptions that happen in the
JavaScript engine during the execution of JSHint.

We now handle the latter too and distinguish both types in the error
message to make the cause of the error more obvious.